backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.54.0
google-generativeai==0.8.3
beautifulsoup4==4.12.2
Pillow==10.1.0
lxml==4.9.3